+++ app/src/main/java/com/countlesswrongs/cryptomonitor/utils/TimeUtils.kt
@@ -2,15 +2,13 @@ package com.countlesswrongs.cryptomonitor.utils
 
 import java.sql.Timestamp
 import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
-import java.util.TimeZone
+import java.util.*
 
-fun convertTimestampToTime(timestamp: Int?): String {
+fun convertTimestampToTime(timestamp: Long?): String {
     if (timestamp == null) {
         return ""
     }
-    val stamp = Timestamp((timestamp * 1_000).toLong())
+    val stamp = Timestamp(timestamp * 1_000)
     val date = Date(stamp.time)
     val pattern = "HH:mm:ss"
     val simpleDateFormat = SimpleDateFormat(pattern, Locale.getDefault())
